Mejoras en algunos scripts

+++ node-alternatives.sh
@@ -1,8 +1,8 @@
-#!/bin/bash
+#!/usr/bin/sudo /bin/bash
 
 BIN="/usr/local/bin"
 
-if [[ $# -ne 1 ]]
+if [[ $# -ne 1 || "$1" == "-h" || "$1" == "--help" ]]
 then
     echo "Usage:   $0 <path to node root>" >&2
     echo "Example: $0 /opt/node-v18.16.1-linux-x64" >&2
@@ -21,11 +21,17 @@ then
     exit 2
 fi
 
-unlink "$BIN/node"
-unlink "$BIN/npm"
+if [[ -L "$BIN/node" ]]
+then
+   unlink "$BIN/node"
+fi
+if [[ -L "$BIN/npm" ]]
+then
+   unlink "$BIN/npm"
+fi
 if [[ -L "$BIN/npx" ]]
 then
-    unlink "$BIN/npx"
+   unlink "$BIN/npx"
 fi
 
 ln -s "$1/bin/node" "$BIN"
